Configuration format should take the name of an OutputWriterFactory, not an OutputWriter
Fixes #60
