+++ jmxtrans2-agent/src/test/resources/config.xml
@@ -33,7 +33,7 @@
     </queries>
 
     <outputWriters>
-        <outputWriter class="org.jmxtrans.core.output.writers.ConsoleOutputWriter"/>
+        <outputWriter class="org.jmxtrans.core.output.writers.ConsoleOutputWriterFactory"/>
     </outputWriters>
 
 </jmxtrans>
