+++ jmxtrans2-core/src/test/java/org/jmxtrans/core/output/writers/ConsoleOutputWriterTest.java
@@ -52,7 +52,7 @@ public class ConsoleOutputWriterTest {
     @Test
     public void factoryCreatesConsoleWriter() {
         Map<String, String> settings = emptyMap();
-        ConsoleOutputWriter consoleOutputWriter = new ConsoleOutputWriter.Factory().create(settings);
+        ConsoleOutputWriter consoleOutputWriter = new ConsoleOutputWriterFactory().create(settings);
         
         assertThat(consoleOutputWriter).isNotNull();
     }
