+++ jmxtrans2-core/src/test/resources/org/jmxtrans/core/config/invalid-configuration.xml
@@ -33,7 +33,7 @@
     </queries>
 
     <outputWriters>
-        <outputWriter class="org.jmxtrans.core.config.DummyOutputWriter"/>
+        <outputWriter class="org.jmxtrans.core.config.DummyOutputWriter$Factory"/>
     </outputWriters>
 
 </jmxtrans>
