+++ jmxtrans2-core/src/test/resources/org/jmxtrans/core/config/no-collection-interval.xml
@@ -30,7 +30,7 @@
     </queries>
 
     <outputWriters>
-        <outputWriter class="org.jmxtrans.core.config.DummyOutputWriter"/>
+        <outputWriter class="org.jmxtrans.core.config.DummyOutputWriter$Factory"/>
     </outputWriters>
 
 </jmxtrans>
