+++ jmxtrans2-spring/src/test/resources/org/jmxtrans/spring/simple-configuration.xml
@@ -46,7 +46,7 @@
     </invocations>
 
     <outputWriters>
-        <outputWriter class="org.jmxtrans.core.output.DevNullOutputWriter"/>
+        <outputWriter class="org.jmxtrans.core.output.DevNullOutputWriter$Factory"/>
     </outputWriters>
 
 </jmxtrans>
