+++ pom.xml
@@ -836,7 +836,7 @@
                 <plugin>
                     <groupId>org.pitest</groupId>
                     <artifactId>pitest-maven</artifactId>
-                    <version>1.1.4</version>
+                    <version>1.1.5</version>
                     <configuration>
                         <mutators>
                             <mutator>DEFAULTS</mutator>
